Remove support for polling via POSIX realtime signals.
Having support for polling via POSIX realtime signals in ivykis
made sense in 2002, since that was the only scalable I/O readiness
notification mechanism supported in the upstream Linux kernel back
then.  But any reasonable Linux install can be assumed to support
epoll these days, so support for POSIX realtime signals can be
removed from ivykis now, as rtsig is a nasty interface, the use of
which should not be encouraged any further.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>