Remove edge-triggered polling methods.
This patches removes the edge-triggered variants of epoll and kqueue,
and makes the level-triggered variants the default poll methods.

There are many reasons for this, the most important of which is
probably the fact that the notification of when a system call issued by
the application on an iv_fd returns -EAGAIN (such a notification is
required by any edge-triggered poll method) is too fragile to implement
in practice.

ivykis does currently provide wrappers for the common system calls
(iv_accept(), iv_connect(), etc) in an attempt to obtain such
notifications, but this only works in a narrow set of situations.
There are many more ways of clearing readiness state on file
descriptors (e.g. getsockopt() SO_ERROR) and it's impossible to catch
them all, and for example for the splice(2) system call, it is
impossible to determine whether an -EAGAIN return means that the input
fd went !POLLIN or the output fd went !POLLOUT.

Removing support for edge triggered poll methods allows some further
simplifications in the ivykis core, which will be done at some point
in the future.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>