Convert the ivykis core to fully level-triggered operation.
This commit converts the ivykis core to expect poll methods to always
re-report every active file descriptor during every poll iteration.  In
other words, it expects poll methods to work in level triggered mode.
(Commit 7231264580cd5c281a6419c76ae2a41518b36d9d ("Remove edge-triggered
polling methods.") documents the reason for switching ivykis to level-
triggered operation.)

What this means in practise is that we can get rid of all expired list
handling and epoch and quotum handling, and simplify things considerably.
In the main loop we simply call the poll method to get a list of all
active file descriptors, and then we invoke the event callbacks for each
of those descriptors, without having to keep any state beyond that point.

The decision on whether to update the kernel's registration of what
bands (IN/OUT/ERR) to deliver events for has been moved from the poll
methods to the ivykis core, since most poll methods had similar logic
for this anyway.  Poll methods are now required to keep
fd->registered_bands up-to-date with a bitmask of which bands
(IN/OUT/ERR) have been registered with the kernel, so that the ivykis
core can decide when to update the kernel registration.

The conditions for disabling or enabling the kernel registration for
a band are as follows:
- Disabling the handler for a band does not immediately disable that
  band with the kernel, the band is disabled only when an event is
  subsequently delivered for that band.
- When enabling the handler for a band that is not currently
  registered with the kernel, we force an update of the kernel
  registration immediately.

The ->reregister_fd() method is now gone, replaced by ->notify_fd(),
which serves mostly the same purpose, but with the twist that the
ivykis core will automatically call ->notify_fd() after ->register_fd()
and before ->unregister_fd().

If the poll method is unable to mask POLLERR (most poll methods are
unable to do so), calling ->notify_fd() with an empty band mask is
required to unregister the file descriptor from its kernel poll
mechanism.  For such mechanisms, all kernel interaction can be left
to ->notify_fd(), and the ->register_fd() and ->unregister-fd() methods
don't need to be implemented at all (for example epoll does this now).

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>