Include splice(2) echo server example application.
Now that ivykis is level-triggered, we can use the splice(2) system
call in ivykis applications.  Include an example echo server that
uses splice(2) to echo data back to the client.

Note that to make this work properly, you'll need to make sure that
your kernel has a patch similar to this one:

	http://marc.info/?l=linux-netdev&m=123113806202824&w=2

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>