Fix oversleeping issue when a signal comes in while polling.
All poll methods had the bug where if the polling system call would
return EINTR, we would restart it, but with the same timeout value as
the previous call, which is wrong.

Change all methods to just return from the poll method if EINTR is
returned, so that the ivykis core will call us again, but then with an
adjusted timeout value.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>