Fix iv_signal deliver versus unregister race.
If an exclusive iv_signal interest is unregistered while it is
active (i.e. it was scheduled to run its callback function, but it
is unregistered before the callback function was run), a signal
event will be lost.

Fix this by having the signal handler set an 'active' flag for the
iv_signal object chosen to handle the signal, having that iv_signal's
ev callback clear this flag, and having iv_signal_unregister() post
the next iv_signal's ev if this one was active and exclusive.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>