iv_wait: Add kill(2) helper that is serialised with respect to wait4(2).
When wait4(2) returns notification of the demise of a process, that
process (which will be in zombie state) is reaped at the same time,
making its process ID available for reuse.  This can lead to a race
condition where one thread calls wait4(2) while another thread tries
to send a signal to the process that the wait4(2) just reaped, which
can cause that signal to end up somewhere unintended.

This commit adds iv_wait_interest_kill(), a helper function that
calls kill(2) while holding iv_wait_lock, so that the call to kill(2)
is serialised with respect to invocations of wait4(2) done by
iv_wait_got_sigchld(), preventing the race described above.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>