iv_wait: Add iv_wait_interest_register_spawn() helper.
Add iv_wait_interest_register_spawn(), which performs a fork() and
then calls iv_wait_interest_register() with the newly created child's
process ID, while keeping the iv_wait_lock held, so that the sequence
is atomic with respect to wait4(2) calls.

This makes sure we catch events for the newly created process even
if they occur before we manage to insert an interest for it into the
iv_wait_interests AVL tree.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>