Build system conversion to autotools.
This commit converts the ivykis build system over to use autotools.
It is based on work by the following individuals, who have all sent
me ivykis autotools patches over the last couple of years:

	Marco Canini <marco.canini@fastwebnet.it>
	Ronald Huizer <r.huizer@xs4all.nl>
	Willem Jan Palenstijn <wjp@usecode.org>
	Balazs Scheidler <bazsi@balabit.hu>

This commit is a bit different from the ivykis autotools patches
submitted so far, in that it tries to stay as close as possible to
the old Makefile setup for now.

In particular, it doesn't build shared libraries yet -- this is
something that we do need to do eventually, and sooner rather than
later, but some more thought will need to go into this before we
can enable building them.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>