Introduce consistent naming of core ivykis fd/task/timer functions.
Rename the following functions:

- INIT_IV_FD -> IV_FD_INIT
- iv_register_fd -> iv_fd_register
- iv_unregister_fd -> iv_fd_unregister
- INIT_IV_TASK -> IV_TASK_INIT
- iv_register_task -> iv_task_register
- iv_unregister_task -> iv_task_unregister
- INIT_IV_TIMER -> IV_TIMER_INIT
- iv_register_timer -> iv_timer_register
- iv_unregister_timer -> iv_timer_unregister

and provide backward compatibility wrappers as inline functions, so
that we get source level backward compatibility (but not binary level).
Mark the backward compat functions with __attribute__((deprecated)) so
that we are warned when the old functions are used.

Suggested by Balazs Scheidler <bazsi@balabit.hu>.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>