This patch adds a new function that returns whether it is possible to poll an FD
On Linux with epoll, you'll never know if an fd can be epolled or not. For
instance regular files cannot be polled, where device files tend to be.
Currently ivykis aborts the whole process if such an error happens.

However there are exceptions, /dev/null is not pollable for instance and
that makes detecting whether a given fd is usable with ivykis a question
that's not easy to decide.

This patch defines a new entry point iv_fd_pollable() which is delegated to
the poll implementation.  It can check whether it's going to be ok to
regsiter that fd.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>