pkg-config files: removed dependent libs and add include location
Dependant libs are not usually published through the .pc file. If the
lib is linked dynamically, it doesn't matter anyway, since the .so will
be linked against the dependencies, no need to do that for the
application using ivykis.

If it is linked statically, then:
  1) there's libtool which could help
    (.la files contain dependencies in this case)
  2) or if it doesn't, the application needs to detection for
     dependencies anyway, and certainly -lrt wouldn't be enough.

I've checked glib, which seems to publish in some cases (depends on
the configure script), and openssl, which doesn't.

Signed-off-by: Tamas Pal <folti@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>