Implement an alternative for TLS variables
Thread-local-storage permits the use of the __thread C language keyword to
indicate that a variable is a per-thread variable. However a lot of
platforms and old GCC/binutils versions have no support for TLS.

This patch introduces an alternative mechanism, which uses TLS with a thin
layer in case it is available, and a pthread_setspecific/getspecific based
implementation otherwise.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>