Fix iv_method_poll local vs. global 'numfds' confusion.
On most poll(2) implementations, POLLHUP/POLLERR are unmaskable, yet
ivykis allows setting all iv_fd handler_* function pointers to NULL,
and so iv_method_poll needs to avoid including iv_fd objects that
don't have any handlers registered in the list of file descriptors it
passes into poll(2).

To this end, it keeps a local set of file descriptors with at least
one non-NULL handler, and a count of the number of file descriptors
in that set, but it ended up confusing this count with the count of
total iv_fd objects registered.

Rename the local 'numfds' variable to 'num_registered_fds' to fix
this.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>