iv_work: return to the main loop every once in a while
iv_work_thread_got_event() can loop forever if the incoming stream of
events is continous, e.g. whenever we finish an item, the next one
is already available. This will cause the ivykis main loop to
never process possibly registered tasks or timers.

This patch changes that function to return to the main loop every
100msecs, even if events would be available.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>