iv_inotify: Rework unregister-from-callback detection.
The current method of keeping a global pointer to the iv_inotify
instance being worked on doesn't work in multithreaded applications.

Use the same method as iv_work uses, which is to stick a pointer to
an on-stack variable into the iv_inotify object being handled in the
event dispatch function, and detect this pointer being non-NULL in
the unregister function and write something through it if that's the
case.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>
Acked-by: Ronald Huizer <r.huizer@xs4all.nl>