iv_fd: Do not abort if trying to replace a non-NULL io handler
The iv_fd_set_handler_{in,out,err} functions check for a condition
where the fd to set a handle on has an old handler registered already,
and abort if they do.

This fails with delayed flushing, as old handlers may not be cleared
yet. Since we do want replacements to work even in this case, the best
course of action is to remove the aborting branch altogether, as they
are not necessary anymore.

Tested-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>