iv_examples(3): Update the NAME section.
The NAME section is used and parsed by tools such as whatis to display
short summaries.  If it does not follow the format expected by the
tool, it will be ignored.

The iv_examples(3) man page is therefore ignored by whatis, which is
unfortunate.  The trivial patch below updates the NAME section, and
makes the page findable via whatis.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>