Trade in -fvisibility= for a linker symbol version script.
Since commit "Use -fvisibility= to hide internal ELF symbols.", on
toolchains that support it, we mark all ELF symbols hidden by default,
and explicitly whitelist symbols that we want to be public by
annotating them with IV_API in the C sources (which resets the
visibility of the symbol to 'default').

This is a step in the right direction, but we can do better: Gergely
Nagy <algernon@balabit.hu> suggested that we start using ELF symbol
versioning for ivykis, and this can actually replace our visibility
tricks entirely.

From now on, if the linker supports Sun style symbol version scripts,
we will use a version script to mark all symbols hidden by default,
and tack a version onto the symbols that we do decide to export.
This allows us to make changes to the API/ABI while still supporting
the older API/ABI from the same shared library, which is a plus if
we'll ever need to do this.

The original patch for this was written by Gergely Nagy
<algernon@balabit.hu>, and he should get most if not all of the
credit for this.

Reported-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>