configure.ac: Test for both pthread_atfork() and pthread_create().
On Linux, linking against libpthread.so.0 doesn't give you
pthread_atfork(), as the latter is defined in
libpthread_nonshared.a, which doesn't get pulled in if
libpthread.so.0 is pulled in indirectly, e.g. via -lrt, but does
get pulled in if you link against -lpthread, as libpthread.so is
a linker script that references both libpthread.so.0 and
libpthread_nonshared.a.

Therefore, we test for pthread_atfork() in AC_SEARCH_LIBS(), to
make sure that libpthread.so really does get pulled in explicitly
via -lpthread rather than implicitly via another shared library
dependency.

However, on FreeBSD, pthread_atfork() is provided by libc.so, and
so, testing for pthread_atfork() in AC_SEARCH_LIBS() doesn't end up
making the other pthreads functions available in the linking stage.

Work around all of this by testing for both pthread_atfork() and
pthread_create() in configure.ac.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>