Introduce iv_fatal(), and route all fatal error conditions through it.
When ivykis detects a fatal error, we shouldn't unconditionally log
the error message to syslog like we do right now.

This commit introduces iv_fatal(), which is a printf-style varargs
function that takes a message for a fatal error, formats it, logs it,
and then aborts the current process.

By default, the message will be logged to syslog as before, unless
iv_set_fatal_msg_handler() was used to set up an alternative error
message log function, in which case it will be logged to that function
instead.

(Original idea from the Balabit ivykis tree.  Suggestion for marking
iv_fatal() with __attribute__((format(printf, 1, 2))) from Gergely
Nagy <algernon@balabit.hu>.)

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>