iv_wait: Implement fairness by list splicing.
When iv_wait_completion() is called, we want it to only process wait
events that were queued before the function was called.  We'll do
this in the same way as iv_task and iv_event do, by moving all wait
events to a local list on entry to the function and then working off
the local list.  This simplifies the locking as well.  If we exit
the function with wait events pending, iv_event guarantees that we'll
be called again at a later time.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>