iv_work: Implement fairness for completion handlers by list splicing.
When iv_work_event() is called, we want it to only invoke work item
completion callbacks for work items that completed before the function
was called.  We'll do this in the (by now) usual way, by moving all
completed work items to a local list on entry to the function and then
working off the local list.  This simplifies the locking as well.  If
we exit the function with completed work items pending, iv_event
guarantees that we'll be called again at a later time.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>