iv_work: Implement fairness for local work items by list splicing.
When iv_work_handle_local() is called, we want it to only process
work items that were registered before the function was called.  We'll
do this in the (by now) usual way, by moving all work items to a local
list on entry to the function and then working off the local list.  If
we exit the function with work items pending, iv_task guarantees that
we'll be called again at a later time.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>