iv_work: Implement fairness for thread work items using sequence numbers.
When iv_work_thread_got_event() is called, we want it to only process
work items that were registered before the function was called.

We can't do this by moving all work items to a local list like it
is done elsewhere in ivykis, since the list of pending work items
is shared between all threads in the thread pool, and so this would
starve other threads of work.

What we can do instead is to (virtually) assign each work item a
sequence number at enqueue time, and when it's time to process work
items, only process work items that have a sequence number less than
or equal to the sequence number of the last work item on the list
when we entered this thread's work processing function.

If we exit iv_work_thread_got_event() with work items pending,
iv_event guarantees that we'll be called again at a later time.

To make things slightly more efficient, we'll make a small change to
the handling of the idle thread list and the idle timer: if we exit
iv_work_thread_got_event() with work items pending, we won't re-add
ourselves to the list of idle threads, and we won't restart the idle
timer.  We'll detect this condition in the next call of the function
by checking whether thr->list is on a list, and if not, we'll skip
deleting ourselves from the list of idle threads and stopping the
idle timer on entry.

The need for thread work item fairness was pointed out by Balazs
Scheidler <bazsi@balabit.hu>.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>