configure.ac: Make pthreads library check work on HP-UX.
On HP-UX, libpthread uses mangled symbol names like on Tru64, but
what's worse, the HP-UX C library provides stub versions of almost
all libpthread functions that return errors instead of doing what
they are supposed to -- and there's no way to detect this at compile
time.

However, there are a number of symbols in libpthread that the C
library doesn't define stubs for, and so we can run AC_SEARCH_LIBS
for one of those functions in libpthread, and if that returns
success, force libpthread to be linked in.  pthread_trace_notify_np
is one such symbol, and we will arbitrarily pick it for this purpose.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>