Work around iv_fd_register_try() being called without handlers.
iv_fd_register_try() was originally implemented so that it would
immediately perform any kernel registrations that it would have had to
perform anyway in the non-synchronous (iv_fd_register()) case.

Meaning that if you would call iv_fd_register_try() on an iv_fd that
has all handlers set to NULL, you would, on most poll methods, get a
success return regardless of the pollability of the file descriptor
you are passing in, since there will be no kernel event registration
done, and thus there is nothing that can return an error.

Having thought about it for a bit, I decided that the current implicit
"you must only call iv_fd_register_try() with at least one non-NULL
handler for the result to be reliable" rule is not a fair one, as it
violates the principle of least surprise.

Therefore, this commit adds some logic to detect the situation where
an fd is passed in that has all handlers set to NULL, and if that is
the case, register it with MASKIN | MASKOUT with the kernel instead.
If that fails, we return an error as usual, while if it succeeds, we
make a note to re-register the fd with the kernel again later, with
the originally intended (null) event mask.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>