iv_wait: Fix potential memory leak introduced by fairness changes.
If an iv_wait_interest is unregistered from its handler callback,
we would fail to free the memory corresponding to any wait events
that were still queued on this pid.

Fix this by always walking the entire pending event list in
iv_wait_completion(), even if we get unregistered halfway through,
as we will just not deliver the events then (but we'll still free
them).

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>