iv_work: Slight optimization for race condition fix.
Instead of doing work in the worker threads directly from the
handler for the iv_event that is used by the main thread to signal a
worker thread, have the iv_event handler schedule an iv_task and do
the work from that iv_task.

The advantage of this is that if we want to quit the work handling
loop because of fairness reasons, and want to have ourselves
rescheduled at some point soon, we can do that by re-registering
our iv_task, which is just a linked list add, and cheaper than an
iv_event_post(), which is a lock/unlock and possibly a linked list
add plus a system call.

To shrink the state space for the various struct work_pool_thread
members, we also introduce the following invariant:

	iv_list_empty(&thr->list) <=>
		iv_task_registered(&thr->work_task) <=>
		!iv_timer_registered(&thr->idle_timer)

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>