configure.ac: Don't use wait4(2) on Solaris.
On Solaris, a negative 'pid' argument to wait4(2) means to wait for
processes that have the absolute value of pid as their process group
ID.  Specifically, calling wait4(2) with a pid value of -1 does not
mean to wait for all child processes, but to wait for processes that
are in process group ID 1, which is not what we want.  The easiest
fix for this is to not use wait4(2), and use waitpid(2) instead.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>