iv_thread_posix: More implementations of get_thread_id().
This commit adds more implementations of iv_thread_posix's version
of get_thread_id().

On all POSIX platforms we currently run on, pthread_t is a scalar
type with sizeof(pthread_t) == sizeof(unsigned long), and so we could
just return the return value of pthread_self() cast to an unsigned
long as a unique thread identifier.

However, the goal of get_thread_id() is to return human-readable
values, and so, using the return value of pthread_self() isn't
always the best idea, since on a number of platforms, the pthread_t
type is used for storing a pointer, and pointers are generally not
very human-readable.

After this commit, we have the following platform-specific
implementations for get_thread_id():

- syscall(__NR_gettid) (sys/syscall.h) on Linux platforms.

- gettid() (process.h) on QNX.

- lwp_gettid() (unistd.h) on DragonFlyBSD.

- void thr_self(long *tid) (sys/ucontext.h + sys/thr.h) on
  FreeBSD.

- thread_t thr_self() (thread.h) (-lthread on AIX 7) on AIX 7,
  SCO OpenServer, Solaris, SCO UnixWare.

If none of these apply, we fall back to returning
(unsigned long)pthread_self() as described above.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>