spinlock.h: Use PTHREAD_PROCESS_SHARED for pthread-backed spinlocks.
iv_signal maintains an AVL tree of registered signal interests,
which is protected from concurrent modifications as well as from
lookup versus modification races by a spinlock.

To be able to reset the disposition of each signal for which there
is an iv_signal object registered back to SIG_DFL in the child
process after fork(), we have to walk the AVL tree from the child
process, but this requires that the child sees a consistent version
of the AVL tree, and that requires that we don't add or remove signal
interests in one thread while calling fork() in another thread in
the parent process.

To guarantee the consistency of the AVL tree data structure across
fork() calls, iv_signal registers atfork handlers which lock the
spinlock in the parent process before the fork() happens, and unlock
the spinlock in both the parent process and the child process after
the fork() completes.

However, for this to work properly, the spinlock that iv_signal uses
must be initialized as a shared spinlock (PTHREAD_PROCESS_SHARED)
instead of as a process private spinlock (PTHREAD_PROCESS_PRIVATE) as
is done currently.

This trips up at least FreeBSD, as FreeBSD's libpthread apparently
has different implementations for process private spinlocks and shared
spinlocks, and trying to lock an unlocked process private spinlock in
a fork()ed child process causes the child to go into an infinite loop.

Fix this by having spinlock.h initialize pthread spinlocks as shared
spinlocks.

This was showing up as an issue on FreeBSD builds of syslog-ng.

Reported-by: Cy Schubert <cy@FreeBSD.org>
Tested-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>