iv_time_win32: Some optimisations.
The result of QueryPerformanceFrequency() doesn't change at runtime,
so we only need to call it once at startup time.

Also, we can avoid doing 64bit-by-32bit divisons and the need for
GetTickCount64() by storing a copy of the last tick count value
divided by 1000 in a global variable and handling subsequent calls
to iv_time_get() as incremental updates of the tick count value.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>