iv_event_raw_posix: Rework eventfd handling.
The Linux glibc eventfd() wrapper function (around the SYS_eventfd{,2}
system calls) returns EINVAL if it is given a nonzero flags argument
and SYS_eventfd2 (which is the variant of SYS_eventfd that takes a flags
argument) isn't implemented, while iv_event_raw_posix was expecting to
get either ENOSYS or success.

Rework iv_event_raw_posix to first try calling eventfd() with the
O_CLOEXEC and O_NONBLOCK flags, and if that fails due to the running
kernel not supporting SYS_eventfd2, fall back to calling eventfd()
with a zero flags argument (which will call SYS_eventfd) and then
setting the two flags on the returned file descriptor by hand.

Also, if the system call numbers for SYS_eventfd or SYS_eventfd2 are
available, call those system calls by hand using syscall(2) instead of
using the glibc eventfd() wrapper function, as the latter insists on
always trying SYS_eventfd2 first, even if that has already been found
to be not implemented in the running kernel and even if the passed-in
flags argument is zero, so doing the system calls by hand allows us to
skip this useless extra guaranteed-to-fail SYS_eventfd2 system call if
that has returned -ENOSYS before.

Reported-by: Dietmar Lingel <dlingel@dit-sec.com>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>