iv_fd_epoll: Don't call epoll_create1() again if it fails with -ENOSYS.
If the running kernel doesn't support epoll_create1(), make a note
of that and don't try calling it again, but instead, immediately fall
back to epoll_create() the next time iv_init() is called.

Also, if possible, call the epoll_create1() system call by hand
(using syscall(2)) so that we avoid introducing an ELF shared library
symbol dependency on GLIBC_2.9.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>