iv_event_raw: Deal with eventfd() EINVAL return.
The Linux glibc eventfd() wrapper function (around the SYS_eventfd{,2}
system calls) returns EINVAL if it is given a nonzero flags argument
and SYS_eventfd2 (which is the variant of SYS_eventfd that takes a flags
argument) isn't implemented, while iv_event_raw was expecting to get
either ENOSYS or success.

Instead of falling back on SYS_eventfd by calling the eventfd() wrapper
again with a zero flags argument and then setting the O_NONBLOCK and
O_CLOEXEC flags by hand, disable use of eventfd on systems that have
SYS_eventfd but not SYS_eventfd2 as a minimally invasive fix for the
stable branches.

Reported by user "Dit" on the Balabit community forums.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>