Turn iv_validate_now() into a no-op.
When support for pthread_[gs]etspecific() was added, 'iv_now' was
changed from being a __thread variable to being a macro dereferencing
the pointer return value of a call to a function in the ivykis core
(commit 5eed214c3e22, "iv_timer: Move per-thread state to struct
iv_state."), with the goal of moving knowledge of how thread-local
state is stored out of the iv.h public header file.

However, this changed the 'call iv_validate_now() / access iv_now'
sequence of operations for accessing the monotonic clock from a
single function call and a thread-local value reference into a
sequence of two function calls, which is suboptimal.

This commit changes iv_validate_now() into an empty define, and adds
the clock validation logic to the function that returns the address
of iv_now (__iv_now_location()), so that all code compiled against a
newer ivykis version will only perform a single function call to
determine the current value of the monotonic clock.

To preserve API compatibility, iv_validate_now() will continue to
exist as a define for as long as we want to keep API/ABI compatibility
(likely forever), and the requirement for ivykis users to call
iv_validate_now() before referencing iv_now has not been dropped by
this commit, to ensure that ivykis users will still work properly when
built and linked against older ivykis versions.

To preserve ABI compatibility for code compiled against an older
ivykis version and dynamically linked against a newer ivykis version,
the old version of iv_validate_now() will continue to be provided as
a symbol exported from libivykis.so.

To prevent ABI problems for code compiled against a newer ivykis
version and dynamically linked against an older ivykis version, the
new version of __iv_now_location() is given a different name, to
prevent code compiled against newer ivykis versions which cause
references to 'iv_now' to validate it if it's out of date from ending
up with a version of __iv_now_location() that doesn't perform this
validation.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>