Further mangle names of the __mutex_*() functions.
QNX 6.5.0 uses the __mutex_{lock,unlock} identifiers for macros
defined in /usr/include/pthread.h, which clashes with our use of
those identifiers.  QNX also uses the variants with a single
underscore, so let's go for triple underscore for the ivykis
identifiers, as the triply underscored versions of the identifiers
don't seem to be in use (yet) on any platforms we run on.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>